\docType{data}

\name{clinical}

\alias{clinical}

\title{Clinical Data of a Cohort of Prostate Cancer Patiens}

\description{The data belong to a cohort of 35 patients with prostate cancer from two different hospitals.
}

\usage{data(clinical)}

\value{
The data.frame "\code{prcomp}" with the following elements: "\code{Hospital}", "\code{Gender}", "\code{Gleason score}", "\code{Ethnicity}", "\code{BMI}", and "\code{Age}".
}

\examples{
data(clinical)

head(clinical)
}

\keyword{datasets}
