
\name{grid-internal}
\alias{grid-internal}
\title{Internal Grid Functions}
\description{
 Internal Grid functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\value{
The return values of these function are used for internal usage.
}
\keyword{ internal }
