% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetRef}
\alias{MetRef}
\title{Nuclear Magnetic Resonance spectra of urine samples}
\format{
A list with three elements:
\describe{
\item{data}{Numeric matrix with 873 rows and 450 columns.}
\item{gender}{Character vector containing donor gender.}
\item{donor}{Character vector containing donor identifiers.}
}
}
\source{
Distributed with the original KODAMA R package.
}
\usage{
data(MetRef)
}
\description{
A longitudinal metabolomics cohort containing 873 urine spectra from 22
healthy donors. Each spectrum contains 450 bins.
}
\references{
Assfalg M, Bertini I, Colangiuli D, et al. (2008)
Evidence of different metabolic phenotypes in humans. PNAS 105:1420-1424.
}
\keyword{datasets}
