% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifolds.R
\name{spirals}
\alias{spirals}
\title{Generate spiral clusters}
\usage{
spirals(n = c(100, 100, 100), sd = c(0, 0, 0))
}
\arguments{
\item{n}{Integer vector giving the number of points in each spiral.}

\item{sd}{Numeric vector giving the radial noise standard deviation for each
spiral.}
}
\value{
A \code{sum(n)} by 2 numeric matrix with columns \code{x} and \code{y}.
}
\description{
Produces two-dimensional spiral clusters using the construction from the
KODAMA R package.
}
\seealso{
\code{\link[=dinisurface]{dinisurface()}}, \code{\link[=helicoid]{helicoid()}}, \code{\link[=swissroll]{swissroll()}}
}
\examples{
dim(spirals(c(50, 50)))
}
