% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifolds.R
\name{dinisurface}
\alias{dinisurface}
\title{Generate Dini's surface}
\usage{
dinisurface(N = 1000)
}
\arguments{
\item{N}{Number of data points.}
}
\value{
An \code{N} by 3 numeric matrix with columns \code{x}, \code{y}, and \code{z}.
}
\description{
Generates points distributed on a Dini surface using the construction from
the KODAMA R package.
}
\examples{
dim(dinisurface(100))
}
\seealso{
\code{\link[=helicoid]{helicoid()}}, \code{\link[=spirals]{spirals()}}, \code{\link[=swissroll]{swissroll()}}
}
