% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifolds.R
\name{helicoid}
\alias{helicoid}
\title{Generate a helicoid surface}
\usage{
helicoid(N = 1000)
}
\arguments{
\item{N}{Number of data points.}
}
\value{
An \code{N} by 3 numeric matrix with columns \code{x}, \code{y}, and \code{z}.
}
\description{
Generates points distributed on a helicoid using the construction from the
KODAMA R package.
}
\examples{
dim(helicoid(100))
}
\seealso{
\code{\link[=dinisurface]{dinisurface()}}, \code{\link[=spirals]{spirals()}}, \code{\link[=swissroll]{swissroll()}}
}
