
\name{grid-internal}
\alias{RQ}
\alias{another}
\alias{knn_Armadillo}
\alias{knnsampling}
\alias{optim.svm.cv}
\alias{samm}
\alias{samm_cpp}
\alias{svm.double.cv}
\alias{svm.kodama}
\alias{vis.KODAMA}
\title{Internal Grid Functions}
\description{
 Internal Grid functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
