% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifolds.R
\name{swissroll}
\alias{swissroll}
\title{Generate a Swiss roll}
\usage{
swissroll(N = 1000)
}
\arguments{
\item{N}{Number of data points.}
}
\value{
An \code{N} by 3 numeric matrix with columns \code{x}, \code{y}, and \code{z}.
}
\description{
Generates the three-dimensional Swiss roll used by the KODAMA R package.
}
\examples{
dim(swissroll(100))
}
\seealso{
\code{\link[=dinisurface]{dinisurface()}}, \code{\link[=helicoid]{helicoid()}}, \code{\link[=spirals]{spirals()}}
}
